+++ apps/backoffice-v2/src/domains/alerts/fetchers.ts
@@ -76,9 +76,13 @@ export const AlertsListSchema = z.array(
   ObjectWithIdSchema.extend({
     dataTimestamp: z.string().datetime(),
     updatedAt: z.string().datetime(),
-    subject: ObjectWithIdSchema.extend({ name: z.string() }),
+    subject: ObjectWithIdSchema.extend({
+      name: z.string(),
+      correlationId: z.string(),
+      type: z.enum(['business', 'counterparty']),
+    }),
     severity: z.enum(AlertSeverities),
-    label: z.string(),
+    correlationId: z.string(),
     alertDetails: z.string(),
     // amountOfTxs: z.number(),
     assignee: ObjectWithIdSchema.extend({
@@ -169,13 +173,13 @@ export const fetchAlertDefinitionByAlertId = async ({ alertId }: { alertId: stri
   return handleZodError(error, alertDefinition);
 };
 
-export const AlertLabelsSchema = z.array(z.string());
+export const AlertCorrelationIdsSchema = z.array(z.string());
 
-export const fetchAlertLabels = async () => {
+export const fetchAlertCorrelationIds = async () => {
   const [alertDefinition, error] = await apiClient({
-    url: `${getOriginUrl(env.VITE_API_URL)}/api/v1/internal/alerts/labels`,
+    url: `${getOriginUrl(env.VITE_API_URL)}/api/v1/internal/alerts/correlationIds`,
     method: Method.GET,
-    schema: AlertLabelsSchema,
+    schema: AlertCorrelationIdsSchema,
   });
 
   return handleZodError(error, alertDefinition);