+++ apps/backoffice-v2/src/lib/blocks/create-blocks-typed/types.ts
@@ -21,6 +21,7 @@ import { CommonWorkflowStates } from '@ballerine/common';
 import { AnyChildren, AnyObject } from '@ballerine/ui';
 import { ColumnDef, TableOptions } from '@tanstack/react-table';
 import { ComponentProps, ReactNode } from 'react';
+import { DataTable } from '@/common/components/organisms/DataTable/DataTable';
 
 export type TBlockCell = {
   type: 'block';
@@ -203,6 +204,11 @@ export type TTableCell = {
   };
 };
 
+export type TDataTableCell = {
+  type: 'dataTable';
+  value: ComponentProps<typeof DataTable>;
+};
+
 export type TParagraphCell = {
   type: 'paragraph';
   value: ReactNode | ReactNode[];
@@ -215,7 +221,7 @@ export type TDialogCell = {
 };
 
 export type TNodeCell = {
-  type: 'nodeCell';
+  type: 'node';
   value: AnyChildren;
 };
 
@@ -236,6 +242,7 @@ export type TCell =
   | TMapCell
   | TCaseCallToActionLegacyCell
   | TTableCell
+  | TDataTableCell
   | TParagraphCell
   | TDialogCell
   | TNodeCell