+++ services/workflows-service/scripts/alerts/generate-alerts.ts
@@ -374,12 +374,12 @@ export const ALERT_DEFINITIONS = {
     },
   },
   TLHAICC: {
-    enabled: true,
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description: `Transaction Limit - Historic Average - Inbound - Inbound transaction exceeds client's historical average`,
     inlineRule: {
       id: 'TLHAICC',
-      fnName: 'evaluateTransactionLimitHistoricAverageInbound',
+      fnName: 'evaluateTransactionAvg',
       subjects: ['counterpartyId'],
       options: {
         transactionDirection: TransactionDirection.inbound,
@@ -390,16 +390,19 @@ export const ALERT_DEFINITIONS = {
         },
         minimumTransactionAmount: 100,
         transactionFactor: 1,
+        customerType: undefined,
+        timeUnit: undefined,
+        timeAmount: undefined,
       },
     },
   },
   TLHAIAPM: {
-    enabled: true,
+    enabled: false,
     defaultSeverity: AlertSeverity.medium,
     description: `Transaction Limit - Historic Average - Inbound - Inbound transaction exceeds client's historical average`,
     inlineRule: {
       id: 'TLHAIAPM',
-      fnName: 'evaluateTransactionLimitHistoricAverageInbound',
+      fnName: 'evaluateTransactionAvg',
       subjects: ['counterpartyId'],
       options: {
         transactionDirection: TransactionDirection.inbound,
@@ -410,6 +413,58 @@ export const ALERT_DEFINITIONS = {
         },
         minimumTransactionAmount: 100,
         transactionFactor: 1,
+        customerType: undefined,
+        timeUnit: undefined,
+        timeAmount: undefined,
+      },
+    },
+  },
+  PGAICT: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `An Credit card inbound transaction value was over the peer group average within a set period of time`,
+    inlineRule: {
+      id: 'PGAICT',
+      fnName: 'evaluateTransactionAvg',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '=',
+        },
+        minimumTransactionAmount: 100,
+        transactionFactor: 2,
+
+        customerType: 'test',
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
+      },
+    },
+  },
+
+  PGAIAPM: {
+    enabled: true,
+    defaultSeverity: AlertSeverity.medium,
+    description: `An non credit card inbound transaction value was over the peer group average within a set period of time`,
+    inlineRule: {
+      id: 'PGAIAPM',
+      fnName: 'evaluateTransactionAvg',
+      subjects: ['counterpartyId'],
+      options: {
+        transactionDirection: TransactionDirection.inbound,
+        minimumCount: 2,
+        paymentMethod: {
+          value: PaymentMethod.credit_card,
+          operator: '!=',
+        },
+        customerType: 'test',
+        minimumTransactionAmount: 100,
+        transactionFactor: 2,
+
+        timeAmount: SEVEN_DAYS,
+        timeUnit: TIME_UNITS.days,
       },
     },
   },
@@ -443,7 +498,7 @@ export const getAlertDefinitionCreateData = (
       ...(dedupeStrategy ?? {}),
     },
     inlineRule,
-    label: id,
+    correlationId: id,
     name: id,
     rulesetId: `set-${id}`,
     description: description,
@@ -461,11 +516,13 @@ export const getAlertDefinitionCreateData = (
 export const generateAlertDefinitions = async (
   prisma: PrismaClient | PrismaTransaction,
   {
-    createdBy = 'SYSTEM',
     project,
+    createdBy = 'SYSTEM',
+    alertsDef = ALERT_DEFINITIONS,
   }: {
     createdBy?: string;
     project: Project;
+    alertsDef?: Partial<typeof ALERT_DEFINITIONS>;
   },
   {
     crossEnvKeyPrefix = undefined,
@@ -474,10 +531,9 @@ export const generateAlertDefinitions = async (
   } = {},
 ) =>
   Promise.all(
-    Object.values(ALERT_DEFINITIONS)
+    Object.values(alertsDef)
       .map(alert => ({
-        label: alert.inlineRule.id,
-        enable: false,
+        correlationId: alert.inlineRule.id,
         ...alert,
       }))
       .filter(alert => alert.enabled)
@@ -489,7 +545,12 @@ export const generateAlertDefinitions = async (
         };
 
         return prisma.alertDefinition.upsert({
-          where: { label_projectId: { label: alertDef.inlineRule.id, projectId: project.id } },
+          where: {
+            correlationId_projectId: {
+              correlationId: alertDef.correlationId,
+              projectId: project.id,
+            },
+          },
           create: getAlertDefinitionCreateData(alertDef, project, createdBy, extraColumns),
           update: getAlertDefinitionCreateData(alertDef, project, createdBy, extraColumns),
           include: {