+++ services/workflows-service/src/alert/alert.controller.external.ts
@@ -55,7 +55,7 @@ export class AlertControllerExternal {
       include: {
         alertDefinition: {
           select: {
-            label: true,
+            correlationId: true,
             description: true,
           },
         },
@@ -96,7 +96,7 @@ export class AlertControllerExternal {
 
       return {
         ...alertWithoutDefinition,
-        label: alertDefinition.label,
+        correlationId: alertDefinition.correlationId,
         assignee: assignee
           ? {
               id: assignee?.id,
@@ -107,11 +107,15 @@ export class AlertControllerExternal {
         alertDetails: alertDefinition.description,
         subject: counterparty.business
           ? {
+              type: 'business',
               id: counterparty.business.id,
               name: counterparty.business.companyName,
+              correlationId: counterparty.business.correlationId,
             }
           : {
+              type: 'counterparty',
               id: counterparty.endUser.id,
+              correlationId: counterparty.endUser.correlationId,
               name: `${counterparty.endUser.firstName} ${counterparty.endUser.lastName}`,
             },
         decision: state,