+++ services/workflows-service/src/alert/alert.service.intg.test.ts
@@ -1,3 +1,4 @@
+import { businessIds } from './../../scripts/generate-end-user';
 import { PrismaService } from '@/prisma/prisma.service';
 import {
   AlertDefinition,
@@ -24,7 +25,6 @@ import {
   ALERT_DEFINITIONS,
   getAlertDefinitionCreateData,
 } from '../../scripts/alerts/generate-alerts';
-import { TExecutionDetails } from './types';
 
 describe('AlertService', () => {
   let prismaService: PrismaService;
@@ -656,6 +656,7 @@ describe('AlertService', () => {
           data: getAlertDefinitionCreateData(
             {
               ...ALERT_DEFINITIONS.TLHAICC,
+              enabled: true,
             },
             project,
           ),
@@ -681,21 +682,32 @@ describe('AlertService', () => {
       });
 
       it('When there are >2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
-        // Arrange
-        await baseTransactionFactory
-          .direction('inbound')
+        const { minimumTransactionAmount } = ALERT_DEFINITIONS.TLHAICC.inlineRule.options;
+
+        const txFactory = transactionFactory
           .paymentMethod(PaymentMethod.credit_card)
-          .withCounterpartyBeneficiary(counteryparty.id)
-          .amount(150)
-          .count(2)
+          .direction(TransactionDirection.inbound)
+          .withCounterpartyBeneficiary(counteryparty.id);
+
+        // Noise
+        await txFactory
+          .paymentMethod(PaymentMethod.apm)
+          .amount(1500)
+          .transactionDate(faker.date.recent(3))
+          .count(1)
           .create();
 
+        // Arrange
+        await txFactory.amount(400).transactionDate(faker.date.past(3)).count(1).create();
+
+        await txFactory.amount(300).transactionDate(faker.date.recent(30)).count(1).create();
+
         await baseTransactionFactory
-          .direction('inbound')
           .paymentMethod(PaymentMethod.credit_card)
-          .withCounterpartyBeneficiary(counteryparty.id)
-          .amount(300)
-          .count(1)
+          .direction(TransactionDirection.inbound)
+          .amount(minimumTransactionAmount + 1)
+          .transactionDate(faker.date.past(2))
+          .count(3)
           .create();
 
         // Act
@@ -711,7 +723,7 @@ describe('AlertService', () => {
       it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
         // Arrange
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.credit_card)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(150)
@@ -719,7 +731,7 @@ describe('AlertService', () => {
           .create();
 
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.credit_card)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(300)
@@ -744,6 +756,7 @@ describe('AlertService', () => {
           data: getAlertDefinitionCreateData(
             {
               ...ALERT_DEFINITIONS.TLHAIAPM,
+              enabled: true,
             },
             project,
           ),
@@ -771,7 +784,7 @@ describe('AlertService', () => {
       it('When there are >2 APM transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
         // Arrange
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.apple_pay)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(150)
@@ -779,7 +792,7 @@ describe('AlertService', () => {
           .create();
 
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.pay_pal)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(300)
@@ -799,7 +812,7 @@ describe('AlertService', () => {
       it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
         // Arrange
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.google_pay)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(150)
@@ -807,7 +820,7 @@ describe('AlertService', () => {
           .create();
 
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.bank_transfer)
           .withCounterpartyBeneficiary(counteryparty.id)
           .amount(300)
@@ -844,7 +857,7 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .withBusinessBeneficiary()
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.credit_card)
           .amount(2)
           .count(ALERT_DEFINITIONS.PAY_HCA_CC.inlineRule.options.amountThreshold + 1)
@@ -866,14 +879,14 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .withBusinessBeneficiary()
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.credit_card)
           .amount(150)
           .count(ALERT_DEFINITIONS.PAY_HCA_CC.inlineRule.options.amountThreshold % 10)
           .create();
 
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .withBusinessBeneficiary()
           .paymentMethod(PaymentMethod.apple_pay)
           .amount(150)
@@ -912,7 +925,7 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .withBusinessBeneficiary()
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.debit_card)
           .amount(2)
           .count(ALERT_DEFINITIONS.PAY_HCA_APM.inlineRule.options.amountThreshold + 1)
@@ -934,18 +947,231 @@ describe('AlertService', () => {
         // Arrange
         await baseTransactionFactory
           .withBusinessBeneficiary()
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .paymentMethod(PaymentMethod.credit_card)
           .amount(150)
           .count(ALERT_DEFINITIONS.PAY_HCA_APM.inlineRule.options.amountThreshold % 10)
           .create();
 
         await baseTransactionFactory
-          .direction('inbound')
+          .direction(TransactionDirection.inbound)
           .withBusinessBeneficiary()
           .paymentMethod(PaymentMethod.apple_pay)
           .amount(150)
           .count(ALERT_DEFINITIONS.PAY_HCA_APM.inlineRule.options.amountThreshold % 10)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+    });
+
+    describe('Rule: PGAICT', () => {
+      let alertDefinition: AlertDefinition;
+      let counteryparty: Counterparty;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              ...ALERT_DEFINITIONS.PGAICT,
+              enabled: true,
+            },
+            project,
+          ),
+        });
+
+        const correlationId = faker.datatype.uuid();
+        counteryparty = await prismaService.counterparty.create({
+          data: {
+            project: { connect: { id: project.id } },
+            correlationId: correlationId,
+            business: {
+              create: {
+                correlationId: correlationId,
+                companyName: faker.company.name(),
+                registrationNumber: faker.datatype.uuid(),
+                mccCode: faker.datatype.number({ min: 1000, max: 9999 }),
+                project: { connect: { id: project.id } },
+                businessType: ALERT_DEFINITIONS.PGAICT.inlineRule.options.customerType,
+              },
+            },
+          },
+        });
+      });
+
+      it('When there are >2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
+        // Noise transactions
+        const { minimumTransactionAmount, transactionFactor, timeAmount } =
+          ALERT_DEFINITIONS.PGAICT.inlineRule.options;
+        await transactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .transactionDate(faker.date.past(2))
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount * transactionFactor * transactionFactor)
+          .count(10)
+          .create();
+
+        // Arrange
+        await transactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .transactionDate(faker.date.recent(timeAmount - 1))
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount + 1)
+          .count(10)
+          .create();
+
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount * transactionFactor + 1)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(counteryparty.id);
+      });
+
+      it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(1)
+          .create();
+
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(0);
+      });
+    });
+
+    describe('Rule: PGAIAPM', () => {
+      let alertDefinition: AlertDefinition;
+      let counteryparty: Counterparty;
+
+      beforeEach(async () => {
+        alertDefinition = await prismaService.alertDefinition.create({
+          data: getAlertDefinitionCreateData(
+            {
+              ...ALERT_DEFINITIONS.PGAIAPM,
+              enabled: true,
+            },
+            project,
+          ),
+        });
+
+        const correlationId = faker.datatype.uuid();
+        counteryparty = await prismaService.counterparty.create({
+          data: {
+            project: { connect: { id: project.id } },
+            correlationId: correlationId,
+            business: {
+              create: {
+                correlationId: correlationId,
+                companyName: faker.company.name(),
+                registrationNumber: faker.datatype.uuid(),
+                mccCode: faker.datatype.number({ min: 1000, max: 9999 }),
+                project: { connect: { id: project.id } },
+                businessType: ALERT_DEFINITIONS.PGAICT.inlineRule.options.customerType,
+              },
+            },
+          },
+        });
+      });
+
+      it('When there are >2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, an alert should be created', async () => {
+        // Noise transactions
+        const { minimumTransactionAmount, transactionFactor, timeAmount } =
+          ALERT_DEFINITIONS.PGAICT.inlineRule.options;
+        await transactionFactory
+          .direction(TransactionDirection.outbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .transactionDate(faker.date.past(2))
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount * transactionFactor * transactionFactor)
+          .count(10)
+          .create();
+
+        await transactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.apm)
+          .transactionDate(faker.date.past(1))
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount + 1)
+          .count(10)
+          .create();
+
+        // Arrange
+        await transactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.apm)
+          .transactionDate(faker.date.recent(timeAmount - 1))
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount + 1)
+          .count(10)
+          .create();
+
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.debit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(minimumTransactionAmount * transactionFactor + 1)
+          .count(1)
+          .create();
+
+        // Act
+        await alertService.checkAllAlerts();
+
+        // Assert
+        const alerts = await prismaService.alert.findMany();
+        expect(alerts).toHaveLength(1);
+        expect(alerts[0]?.alertDefinitionId).toEqual(alertDefinition.id);
+        expect(alerts[0]?.counterpartyId).toEqual(counteryparty.id);
+      });
+
+      it('When there are 2 credit card transactions with >100 base amount and one transaction exceeds the average of all credit card transactions, no alert should be created', async () => {
+        // Arrange
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(150)
+          .count(1)
+          .create();
+
+        await baseTransactionFactory
+          .direction(TransactionDirection.inbound)
+          .paymentMethod(PaymentMethod.credit_card)
+          .withCounterpartyBeneficiary(counteryparty.id)
+          .amount(300)
+          .count(1)
           .create();
 
         // Act