+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # kyb-app
 
+## 0.3.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.5
+  - @ballerine/workflow-browser-sdk@0.6.8
+
 ## 0.3.7
 
 ### Patch Changes